Se agrego la venta completa , se agregaron las salidas en cada venta, se agrego la funcion de hacer un corte de cajaasimismo con autorizacion de adminsitrador para aceptar el corte, se agrego el cancelar venta, se agrego el cerrar sesion en la pantalla de caja, ya no se puede modificar el precio al querer modificar un producto, ya no se puede eliminar un producto si este ha estando en la venta, el corte de caja hace un archivo excel, asimismo que al iniciar la sesion de cajero crea un respaldo de las ventas

+++ vista/src/main/java/ui/InventarioBean.java
@@ -117,11 +117,14 @@ public class InventarioBean implements Serializable {
             msgWarn("La cantidad debe ser mayor a cero.");
             return;
         }
+        if(cantidadOperacion > dto.getStock()){
+            msgWarn("La cantidad de salida supera al Stock actual, ingrese otra cantidad");
+        }
         try{
             Producto p = new Producto();
             p.setId(dto.getIdProducto());
             p.setNombre(dto.getNombreProducto());
-            int nuevoStock = dto.getStock() + cantidadOperacion;
+            int nuevoStock = dto.getStock() - cantidadOperacion;
             facade.actualizarStock(dto.getIdProducto(), nuevoStock);
             facade.registrarMovimientoSeguro(usuarioLogeado,p, "Salida", cantidadOperacion);
             dto.setStock(nuevoStock);
@@ -222,7 +225,7 @@ public class InventarioBean implements Serializable {
     }
 
     public void prepararEliminar() {
-
+        System.out.println("Nombre del producto: " + seleccionado.getNombreProducto());
         if (seleccionado == null) {
             msgWarn("Seleccione un producto primero");
             return;
@@ -232,15 +235,17 @@ public class InventarioBean implements Serializable {
     }
 
     public void eliminarProducto() {
+        System.out.println("Entre al eliminarProducto");
         try {
             if (seleccionado == null || seleccionado.getIdProducto() <= 0) {
                 return;
             }
-
             int idProducto = seleccionado.getIdProducto();
 
-
-
+            if(facade.existeenVenta(idProducto)){
+                msgWarn("El producto no se puede eliminar, ya que esta ligado a una o mas ventas");
+                return;
+            }
             facade.eliminarRegistroInventario(idProducto);
             facade.eliminarProducto(idProducto);
             seleccionado = null;
